Rename concretize.rs to concretizer.rs to match Lean version
